refactor: 캐릭터 수정 컴포넌트 분리
